Reverse engineered the FTP classes, now compiles again
After my laptop was stolen I realised I didn't have a bunch of the files backed up.
I had the generated PHP on a server though, and was able to reverse engineer it mostly.
The files are not commented / documented, sorry...

It now compiles, though javascript is not running on my firefox for some reason